fixup! [ADD] sms_twilio: new SMS provider

+++ addons/sms_twilio/wizard/manage_connection_wizard.py
@@ -21,9 +21,6 @@ class SmsTwilioManageConnectionWizard(models.TransientModel):
     sms_twilio_number_ids = fields.One2many(related='company_id.sms_twilio_number_ids', readonly=False)
     sms_twilio_to_number = fields.Char("To Number")
 
-    sms_twilio_success_msg = fields.Char("Success message", readonly=True)
-    sms_twilio_error_msg = fields.Char("Error message", readonly=True)
-
     def reload_numbers(self):
         """Fetch the available numbers from Twilio account"""
         self.company_id._assert_twilio_sid()
@@ -35,9 +32,19 @@ class SmsTwilioManageConnectionWizard(models.TransientModel):
             )
         except requests.exceptions.RequestException as e:
             _logger.warning('Twilio SMS API error: %s', str(e))
-            return self.company_id._action_sms_twilio_open_manage_connection_wizard(self)
+            return self._display_notification(
+                notif_type='danger',
+                message= _("An error occurred while fetching the numbers."),
+            )
 
         json_response = response.json()
+        if not response.ok:
+            _logger.warning('Twilio SMS API error: %s', json_response.get('code'))
+            return self._display_notification(
+                notif_type='danger',
+                message=_("Error: %s", json_response.get('message')),
+            )
+
         self.sms_twilio_number_ids.unlink()
         for twilio_number in json_response.get('incoming_phone_numbers', []):
             country_code = get_country_code_from_phone(twilio_number.get('phone_number'))
@@ -68,17 +75,33 @@ class SmsTwilioManageConnectionWizard(models.TransientModel):
             active_id=temp_partner,
         ).create({'body': _("This is a test SMS from Odoo")})
         message_notif = composer._action_send_sms().notification_ids[0]
+
+        res_error = False
         if not message_notif.failure_type:
-            self.sms_twilio_success_msg = _("The SMS has been sent from %s", get_twilio_from_number(self.company_id, self.sms_twilio_to_number).display_name)
-            self.sms_twilio_error_msg = False
+            res_msg = _("The SMS has been sent from %s", get_twilio_from_number(self.company_id, self.sms_twilio_to_number).display_name)
         else:
-            self.sms_twilio_success_msg = False
+            res_error = True
             if message_notif.failure_type != "unknown":
-                self.sms_twilio_error_msg = dict(message_notif._fields["failure_type"].selection).get(message_notif.failure_type)  # Translated error msg
+                res_msg = dict(message_notif._fields["failure_type"].selection).get(message_notif.failure_type)  # Translated error msg
             else:
-                self.sms_twilio_error_msg = message_notif.failure_reason  # Not translated, coming from Twilio
+                res_msg = _("Error: %s", message_notif.failure_reason)  # Not translated, coming from Twilio
         temp_partner.unlink()
-        return self.company_id._action_sms_twilio_open_manage_connection_wizard(self)
+        return self._display_notification(
+            notif_type='danger' if res_error else 'success',
+            message=res_msg,
+        )
 
     def action_save(self):
         return {'type': 'ir.actions.act_window_close'}
+
+    def _display_notification(self, notif_type, message):
+        return {
+            'type': 'ir.actions.client',
+            'tag': 'display_notification',
+            'params': {
+                'title': _("Twilio SMS"),
+                'message': message,
+                'type': notif_type,
+                'sticky': False,
+            }
+        }
